refs code #3 next days weather

+++ app/src/main/java/com/duycuong/weather/utils/TimeUtils.java
@@ -8,7 +8,7 @@ import java.util.Locale;
  */
 
 public class TimeUtils {
-    public static String covertTime(String format, int time) {
+    public static String covertTime(String format, long time) {
         SimpleDateFormat simpleDateFormat =
                 new SimpleDateFormat(format, Locale.getDefault());
         return simpleDateFormat.format(time * 1000);
